requirements.txt
# Core Libraries
spotipy
fastapi
uvicorn
requests
numpy
pandas
nltk

# LLM Processing
groq

# Collaborative Filtering / Matrix Factorization
scikit-surprise

# Optional Utilities
matplotlib
python-dotenv
orjson
